+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added game version ranged swing reset specificity.
 - Added channeled spell swing reset.
 
+### Fixed
+- Fix preventSwingReset flag. Prevent flag from being stuck to true after channeling a spell.
+
 ## [1.2.0] - 2022-08-30
 
 ### Added